Backend/vedya/api/views.py
from django.http import JsonResponse, HttpResponse
from django.views.decorators.csrf import csrf_exempt
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from twilio.twiml.messaging_response import MessagingResponse
import json

# Import needed services and models here

@csrf_exempt
def twilio_webhook(request):
    """Endpoint for handling incoming WhatsApp messages from Twilio"""
    if request.method == 'POST':
        # Extract incoming message details
        incoming_msg = request.POST.get('Body', '').strip()
        sender = request.POST.get('From', '')
        
        # Create a response
        resp = MessagingResponse()
        
        # TODO: Process message with Patient AI Agent
        # This will be replaced with actual AI agent call
        resp.message(f"Thank you for your message: {incoming_msg}. Our AI assistant is processing your request.")
        
        return HttpResponse(str(resp))
    
    return HttpResponse(status=405)

@api_view(['GET', 'POST'])
def doctor_list(request):
    """List all doctors or create a new doctor"""
    if request.method == 'GET':
        # TODO: Fetch doctors from database
        doctors = []  # Replace with actual database query
        return Response(doctors)
    
    elif request.method == 'POST':
        # TODO: Create new doctor in database
        return Response({'message': 'Doctor created'}, status=status.HTTP_201_CREATED)

@api_view(['GET', 'POST'])
def patient_list(request):
    """List all patients or create a new patient"""
    if request.method == 'GET':
        # TODO: Fetch patients from database
        patients = []  # Replace with actual database query
        return Response(patients)
    
    elif request.method == 'POST':
        # TODO: Create new patient in database
        return Response({'message': 'Patient created'}, status=status.HTTP_201_CREATED)

@api_view(['GET', 'POST'])
def appointment_list(request):
    """List all appointments or create a new appointment"""
    if request.method == 'GET':
        # TODO: Fetch appointments from database
        appointments = []  # Replace with actual database query
        return Response(appointments)
    
    elif request.method == 'POST':
        # TODO: Create new appointment in database
        return Response({'message': 'Appointment created'}, status=status.HTTP_201_CREATED)

@api_view(['GET', 'PUT', 'DELETE'])
def appointment_detail(request, appointment_id):
    """Retrieve, update or delete an appointment"""
    # TODO: Get appointment by ID from database
    appointment = None  # Replace with actual database query
    
    if appointment is None:
        return Response(status=status.HTTP_404_NOT_FOUND)
    
    if request.method == 'GET':
        # TODO: Serialize and return appointment
        return Response({"id": appointment_id})
    
    elif request.method == 'PUT':
        # TODO: Update appointment in database
        return Response({'message': 'Appointment updated'})
    
    elif request.method == 'DELETE':
        # TODO: Delete appointment from database
        return Response(status=status.HTTP_204_NO_CONTENT)
