Backend/requirements.txt
django==5.0.0
django-rest-framework==0.1.0
django-cors-headers==4.3.1
pymongo==4.6.1
twilio==8.10.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.0
langchaingraph==0.0.20
pydantic==2.5.2
asyncio==3.4.3
